Partial fix to spectrogram misalignment
Now the top and bottom are aligned, and there is proper clipping in log plot. The bottom element should either be included using rounding up or similar, or the bottoms should be pulled down so all space is used.

However, the main spectrograph cells don't line up with the single spectrum plot or the mouse over tooltip